fix: old component isolation
